requirements.txt
flask
requests
datetime
jsonify
python-dotenv
psycopg2-binary
